+++ R/utils.R
@@ -5,3 +5,8 @@ error_class <- function(class) {
     paste0("gsm.digitpref-error-", class)
   )
 }
+
+nameColumns <- function(dfData, ...) {
+  all_names <- colnames(dplyr::select(dfData, ...))
+  glue::glue("{all_names}")
+}