+++ tests/testthat/test-plotDigitCounts.R
@@ -1,10 +1,10 @@
-test_that("stackedDigitChart produces an expected chart", {
+test_that("plotDigitCounts produces an expected chart", {
   set.seed(42)
   sample_data <- data.frame(
     siteID = sample(c("siteA", "siteB", "siteC"), 1000, replace = TRUE),
     onesPlace = sample(0:9, 1000, replace = TRUE)
   )
-  test_result <- stackedDigitChart(sample_data, siteID, onesPlace)
+  test_result <- plotDigitCounts(sample_data, siteID, onesPlace)
   expect_s3_class(test_result, "ggplot")
   expect_identical(as.character(test_result$labels$x), "siteID")
   expect_identical(as.character(test_result$labels$y), "Frequency")
@@ -14,18 +14,18 @@ test_that("stackedDigitChart produces an expected chart", {
     "Digit Distribution for onesPlace of Data by siteID"
   )
   vdiffr::expect_doppelganger(
-    title = "stackedDigitChart default",
+    title = "plotDigitCounts default",
     fig = test_result
   )
 })
 
-test_that("stackedDigitChart accepts supplied group label", {
+test_that("plotDigitCounts accepts supplied group label", {
   set.seed(42)
   sample_data <- data.frame(
     siteID = sample(c("siteA", "siteB", "siteC"), 1000, replace = TRUE),
     onesPlace = sample(0:9, 1000, replace = TRUE)
   )
-  test_result <- stackedDigitChart(
+  test_result <- plotDigitCounts(
     sample_data,
     siteID,
     onesPlace,
@@ -39,13 +39,13 @@ test_that("stackedDigitChart accepts supplied group label", {
   )
 })
 
-test_that("stackedDigitChart accepts supplied place label", {
+test_that("plotDigitCounts accepts supplied place label", {
   set.seed(42)
   sample_data <- data.frame(
     siteID = sample(c("siteA", "siteB", "siteC"), 1000, replace = TRUE),
     onesPlace = sample(0:9, 1000, replace = TRUE)
   )
-  test_result <- stackedDigitChart(
+  test_result <- plotDigitCounts(
     sample_data,
     siteID,
     onesPlace,
@@ -58,13 +58,13 @@ test_that("stackedDigitChart accepts supplied place label", {
   )
 })
 
-test_that("stackedDigitChart accepts title override", {
+test_that("plotDigitCounts accepts title override", {
   set.seed(42)
   sample_data <- data.frame(
     siteID = sample(c("siteA", "siteB", "siteC"), 1000, replace = TRUE),
     onesPlace = sample(0:9, 1000, replace = TRUE)
   )
-  test_result <- stackedDigitChart(
+  test_result <- plotDigitCounts(
     sample_data,
     siteID,
     onesPlace,
@@ -79,15 +79,15 @@ test_that("stackedDigitChart accepts title override", {
   )
 })
 
-test_that("stackedDigitChart accepts scaleDigitPalette override", {
+test_that("plotDigitCounts accepts scaleDigitPalette override", {
   set.seed(42)
   sample_data <- data.frame(
     siteID = sample(c("siteA", "siteB", "siteC"), 1000, replace = TRUE),
     onesPlace = sample(0:9, 1000, replace = TRUE)
   )
   vdiffr::expect_doppelganger(
-    title = "stackedDigitChart no palette",
-    fig = stackedDigitChart(
+    title = "plotDigitCounts no palette",
+    fig = plotDigitCounts(
       sample_data,
       siteID,
       onesPlace,
@@ -95,8 +95,8 @@ test_that("stackedDigitChart accepts scaleDigitPalette override", {
     )
   )
   vdiffr::expect_doppelganger(
-    title = "stackedDigitChart paired palette",
-    fig = stackedDigitChart(
+    title = "plotDigitCounts paired palette",
+    fig = plotDigitCounts(
       sample_data,
       siteID,
       onesPlace,
@@ -105,15 +105,15 @@ test_that("stackedDigitChart accepts scaleDigitPalette override", {
   )
 })
 
-test_that("stackedDigitChart accepts theme override", {
+test_that("plotDigitCounts accepts theme override", {
   set.seed(42)
   sample_data <- data.frame(
     siteID = sample(c("siteA", "siteB", "siteC"), 1000, replace = TRUE),
     onesPlace = sample(0:9, 1000, replace = TRUE)
   )
   vdiffr::expect_doppelganger(
-    title = "stackedDigitChart no theme",
-    fig = stackedDigitChart(
+    title = "plotDigitCounts no theme",
+    fig = plotDigitCounts(
       sample_data,
       siteID,
       onesPlace,
@@ -121,8 +121,8 @@ test_that("stackedDigitChart accepts theme override", {
     )
   )
   vdiffr::expect_doppelganger(
-    title = "stackedDigitChart theme_minimal",
-    fig = stackedDigitChart(
+    title = "plotDigitCounts theme_minimal",
+    fig = plotDigitCounts(
       sample_data,
       siteID,
       onesPlace,